requirements.txt
streamlit
opencv-python-headless
numpy
pandas
easyocr
Pillow
zipfile36
shutil
